UserProfile, UserPost were combined into UserPage
